Move all code to `gocv` package.
