+++ gocv/gocv.go
@@ -1,4 +1,4 @@
-package gcv_core
+package gocv
 
 // #cgo CXXFLAGS: -std=c++11
 // #cgo darwin pkg-config: opencv
