--- opencv2/gcv_calib3d/gcv_calib3d.go
@@ -1,5 +1,0 @@
-package gcv_calib3d
-
-// #cgo CXXFLAGS: -std=c++11
-// #cgo darwin pkg-config: opencv
-import "C"
