--- opencv2/gcv_calib3d/gcv_calib3d.swigcxx
@@ -1,7 +1,0 @@
-%module gcv_calib3d
-
-%{
-#include "gcv_calib3d.hpp"
-%}
-
-%include "gcv_calib3d.hpp"
