--- opencv2/opencv2.go
@@ -1,1 +1,0 @@
-package opencv2
